+++ CHANGELOG.md
@@ -36,6 +36,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - a new view that show all completed conversion projects that are voluntary at
   `/projects/all/in-progress/voluntary`
 - basic telemetery is being sent to DfE Application Insights
+- Add a link to the companies house information page from the project
+  information tab
 
 ### Fixed